Split out S3PresignedUrls as a separate class
