+++ storage/src/main/scala/weco/storage/transfer/azure/AzureTransfer.scala
@@ -4,16 +4,12 @@ import java.io.{ByteArrayInputStream, InputStream}
 import java.net.URL
 import com.amazonaws.services.s3.AmazonS3
 import com.azure.storage.blob.BlobServiceClient
-import com.azure.storage.blob.models.{
-  BlobRange,
-  BlobStorageException,
-  BlockListType
-}
+import com.azure.storage.blob.models.{BlobRange, BlobStorageException, BlockListType}
 import com.azure.storage.blob.specialized.BlockBlobClient
 import grizzled.slf4j.Logging
 import org.apache.commons.io.IOUtils
 import weco.storage.models.ByteRange
-import weco.storage.services.s3.{S3RangedReader, S3Uploader}
+import weco.storage.services.s3.{S3PresignedUrls, S3RangedReader}
 import weco.storage.transfer._
 import weco.storage.{Identified, NotFoundError, RetryableError, StoreWriteError}
 import weco.storage.azure.AzureBlobLocation
@@ -21,11 +17,7 @@ import weco.storage.s3.S3ObjectLocation
 import weco.storage.services.azure.AzureSizeFinder
 import weco.storage.store.azure.AzureStreamStore
 import weco.storage.store.s3.S3StreamStore
-import weco.storage.transfer.{
-  TransferNoOp,
-  TransferOverwriteFailure,
-  TransferSourceFailure
-}
+import weco.storage.transfer.{TransferNoOp, TransferOverwriteFailure, TransferSourceFailure}
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
@@ -321,8 +313,7 @@ class AzurePutBlockTransfer(
   }
 }
 
-class AzurePutBlockFromUrlTransfer(s3Uploader: S3Uploader,
-                                   azureSizeFinder: AzureSizeFinder,
+class AzurePutBlockFromUrlTransfer(azureSizeFinder: AzureSizeFinder,
                                    blockTransfer: AzurePutBlockTransfer)(
   signedUrlValidity: FiniteDuration,
   val blockSize: Long)(
@@ -331,11 +322,13 @@ class AzurePutBlockFromUrlTransfer(s3Uploader: S3Uploader,
   val blobServiceClient: BlobServiceClient
 ) extends AzureTransfer[URL] {
 
+  private val s3Presigner = new S3PresignedUrls()
+
   override protected def getContext(
     src: SourceS3Object,
     dst: AzureBlobLocation
   ): Either[TransferSourceFailure[SourceS3Object, AzureBlobLocation], URL] =
-    s3Uploader
+    s3Presigner
       .getPresignedGetURL(
         location = src.location,
         expiryLength = signedUrlValidity)
@@ -399,7 +392,6 @@ object AzurePutBlockFromUrlTransfer {
   def apply(signedUrlValidity: FiniteDuration, blockSize: Long)(
     implicit s3Client: AmazonS3,
     blobServiceClient: BlobServiceClient) = {
-    val s3Uploader = new S3Uploader()
 
     // In the actual replicator, if there's an object in S3 and an object in Azure,
     // assume they're both the same.  The verifier will validate the checksum later.
@@ -422,7 +414,6 @@ object AzurePutBlockFromUrlTransfer {
     // it for a handful of keys.
     val blockTransfer = new AzurePutBlockTransfer(blockSize = blockSize)
     new AzurePutBlockFromUrlTransfer(
-      s3Uploader,
       azureSizeFinder,
       blockTransfer)(signedUrlValidity, blockSize)
   }
